ts lib set up
