+++ js/src/artifacts/index.ts
@@ -1,0 +1,5 @@
+import OpenBankingDomesticCircuit from "./circuit/openbanking_domestic.json";
+
+export {
+    OpenBankingDomesticCircuit
+};
