+++ js/src/constants.ts
@@ -1,0 +1,3 @@
+export const MAX_AMOUNT_LENGTH = 10;
+export const MAX_JWT_SIZE = 1536;
+export const MAX_PAYLOAD_SIZE = 1024;
