+++ lib/Nargo.toml
@@ -4,7 +4,7 @@ type = "lib"
 compiler_version = ">=1.0.0"
 
 [dependencies]
-aztec = { git="https://github.com/AztecProtocol/aztec-packages/", tag="aztec-packages-v0.72.1", directory="noir-projects/aztec-nr/aztec" }
+# aztec = { git="https://github.com/AztecProtocol/aztec-packages/", tag="aztec-packages-v0.72.1", directory="noir-projects/aztec-nr/aztec" }
 bignum = {tag = "v0.5.0", git = "https://github.com/noir-lang/noir-bignum"}
 json_parser = { tag = "main", git = "https://github.com/noir-lang/noir_json_parser"}
 rsa_pss = { git = "https://github.com/jp4g/noir_rsa_pss", tag = "feat/nargo-1.0.0", directory = "lib" }
